+++ jave-core-test/src/test/java/ws/schild/jave/ScreenExtractorTest.java
@@ -22,6 +22,7 @@ public class ScreenExtractorTest extends AMediaTest{
     
     /**
      * Test of render method, of class ScreenExtractor.
+     * @throws java.lang.Exception
      */
     @Test
     public void testRenderImages_01() throws Exception {
@@ -52,6 +53,7 @@ public class ScreenExtractorTest extends AMediaTest{
 
     /**
      * Test of render method, of class ScreenExtractor.
+     * @throws java.lang.Exception
      */
     @Test
     public void testRenderImages_02() throws Exception {
@@ -82,6 +84,7 @@ public class ScreenExtractorTest extends AMediaTest{
 
     /**
      * Test of render method, of class ScreenExtractor.
+     * @throws java.lang.Exception
      */
     @Test
     public void testRenderImages_03() throws Exception {
@@ -112,6 +115,7 @@ public class ScreenExtractorTest extends AMediaTest{
     
     /**
      * Test of render method, of class ScreenExtractor.
+     * @throws java.lang.Exception
      */
     @Test
     public void testRenderImage_01() throws Exception {
@@ -135,6 +139,7 @@ public class ScreenExtractorTest extends AMediaTest{
     
     /**
      * Test of render method, of class ScreenExtractor.
+     * @throws java.lang.Exception
      */
     @Test
     public void testRenderImage_02() throws Exception {
@@ -154,5 +159,51 @@ public class ScreenExtractorTest extends AMediaTest{
         instance.render(multimediaObject, width, height, seconds, target, quality);
         assertTrue(target.exists(), "Output file missing");
     }
+
+    /**
+     * Test of render method, of class ScreenExtractor.
+     * @throws java.lang.Exception
+     */
+    @Test
+    public void testRenderOneImage00() throws Exception {
+        System.out.println("render one image 00");
+        File source = new File(getResourceSourcePath(), "zelda first commercial.mpeg");
+        File target = new File(getResourceTargetPath(), "RenderOneImage00.jpg");
+        if (target.exists())
+        {
+            target.delete();
+        }
+        MultimediaObject multimediaObject = new MultimediaObject(source);
+        int width = -1;
+        int height = -1;
+        int seconds = 15000;
+        int quality = 1;
+        ScreenExtractor instance = new ScreenExtractor();
+        instance.renderOneImage(multimediaObject, width, height, seconds, target, quality);
+        assertTrue(target.exists(), "Output file missing");
+    }
     
+    
+    /**
+     * Test of render method, of class ScreenExtractor.
+     * @throws java.lang.Exception
+     */
+    @Test
+    public void testRenderOneImage01() throws Exception {
+        System.out.println("render one image 01");
+        File source = new File(getResourceSourcePath(), "zelda first commercial.mpeg");
+        File target = new File(getResourceTargetPath(), "RenderOneImage01.jpg");
+        if (target.exists())
+        {
+            target.delete();
+        }
+        MultimediaObject multimediaObject = new MultimediaObject(source);
+        int width = 60;
+        int height = 60;
+        int seconds = 15000;
+        int quality = 1;
+        ScreenExtractor instance = new ScreenExtractor();
+        instance.renderOneImage(multimediaObject, width, height, seconds, target, quality);
+        assertTrue(target.exists(), "Output file missing");
+    }
 }